restructuring for automation
Mechanism in place to automate setup of software tools. VS Code is one example in this commit; installing for both OSX and Windows, as well as using the same key bindings across both.

TODO: keybindings only work if power toys are installed and Windows key is mapped to control key. This is not automatically done via the script yet.